refactor(pi): share one terminal-control stripper
PiConversation and the terminal renderer's Ansi helper carried byte-identical
control-sequence strippers. Extract a single `orca.util.TerminalControl` in
tools and have both delegate to it.
